Update Issue REC-25: Add authorization

+++ RecipeApi/Service/RecipeInstructionService.cs
@@ -1,8 +1,11 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using AutoMapper;
+using Microsoft.AspNetCore.Authorization;
+using RecipeApi.Authentication;
 using RecipeApi.Database;
 using RecipeApi.Exceptions;
 using RecipeApi.IService;
@@ -15,19 +18,31 @@ namespace RecipeApi.Service
         private readonly RecipeDbContext _dbContext;
         private readonly IMapper _mapper;
         private readonly ILogger _logger;
+        private readonly IAuthorizationService _authorizationService;
+        private readonly IUserContentService _userContentService;
 
-        public RecipeInstructionService(RecipeDbContext context, IMapper mapper, ILogger<RecipeInstructionService> logger)
+
+        public RecipeInstructionService(RecipeDbContext context,
+        IMapper mapper,
+        ILogger<RecipeInstructionService> logger,
+        IAuthorizationService authorizationService,
+        IUserContentService userContentService)
         {
             _dbContext = context;
             _mapper = mapper;
             _logger = logger;
+            _authorizationService = authorizationService;
+            _userContentService = userContentService;
         }
 
         public int AddRecipeInstruction(CreateRecipeInstructionToExistingRecipeDto instructionDto)
         {
-            if(instructionDto is null) throw new BadRequestException("Empty instruction data");
+            if (instructionDto is null) throw new BadRequestException("Empty instruction data");
 
             var instruction = _mapper.Map<RecipeInstruction>(instructionDto);
+
+            checkAuthorization(_userContentService.User, instructionDto.RecipeId);
+
             _dbContext.RecipeInstructions.Add(instruction);
             _dbContext.SaveChanges();
 
@@ -41,6 +56,8 @@ namespace RecipeApi.Service
             var instruction = _dbContext.RecipeInstructions.FirstOrDefault(i => i.Id == id);
 
             if (instruction is null) throw new NotFoundException("Instruction not found");
+
+            checkAuthorization(_userContentService.User, instruction.RecipeId);
 
             _dbContext.RecipeInstructions.Remove(instruction);
             _dbContext.SaveChanges();
@@ -58,8 +75,8 @@ namespace RecipeApi.Service
         public IEnumerable<RecipeInstructionDto> GetRecipeInstructionsByRecipeId(int recipeId)
         {
             var instructions = _dbContext.RecipeInstructions.Where(i => i.RecipeId == recipeId).OrderBy(i => i.Order).ToList();
-            
-            if (instructions.Count==0) throw new NotFoundException("Instructions not found");
+
+            if (instructions.Count == 0) throw new NotFoundException("Instructions not found");
 
             return _mapper.Map<IEnumerable<RecipeInstructionDto>>(instructions);
         }
@@ -67,15 +84,34 @@ namespace RecipeApi.Service
 
         public void UpdateRecipeInstruction(int id, UpdateRecipeInstructionDto instruction)
         {
+
             var instructionToUpdate = _dbContext.RecipeInstructions.FirstOrDefault(i => i.Id == id);
 
             if (instructionToUpdate is null) throw new NotFoundException("Instruction not found");
+
+            checkAuthorization(_userContentService.User, instructionToUpdate.RecipeId);
 
             instructionToUpdate.Instruction = instruction.Instruction;
             instructionToUpdate.Order = instruction.Order;
 
             _dbContext.RecipeInstructions.Update(instructionToUpdate);
             _dbContext.SaveChanges();
+        }
+        private void checkAuthorization(ClaimsPrincipal user, int recipeId)
+        {
+            var recipe = _dbContext.Recipes.FirstOrDefault(r => r.Id == recipeId);
+            if (recipe is null) throw new NotFoundException("Recipe not found");
+
+            var authorizationResult = _authorizationService.AuthorizeAsync(user, recipe,
+                new SameAuthorRequirement()).Result;
+            if (!authorizationResult.Succeeded)
+            {
+                throw new ForbidException("You are not authorized to perform this action");
+            }
+            else
+            {
+                recipe.UpdatedAt = DateTime.Now.ToUniversalTime();
+            }
         }
     }
 }
